Move ToString to IIgniteTuple
