+++ modules/platforms/dotnet/Apache.Ignite/Table/IIgniteTuple.cs
@@ -128,5 +128,23 @@ namespace Apache.Ignite.Table
 
             return true;
         }
+
+        /// <summary>
+        /// Converts tuple to string.
+        /// </summary>
+        /// <param name="tuple">Tuple.</param>
+        /// <returns>String representation.</returns>
+        public static string ToString(IIgniteTuple tuple)
+        {
+            IgniteArgumentCheck.NotNull(tuple, nameof(tuple));
+            var builder = new IgniteToStringBuilder(tuple.GetType());
+
+            for (var i = 0; i < tuple.FieldCount; i++)
+            {
+                builder.Append(tuple[i], tuple.GetName(i));
+            }
+
+            return builder.Build();
+        }
     }
 }
